Populating sub directories correctly in Shared Directory (#15982) (#15991)
## Description

Populating sub directories correctly in Shared Directory. When we
serialize a new directory on attach and record the csn which is -1 for a
newly create one, then on load, we load with -1 csn for that. Now there
is no create op for this subdir. So when a new subdir is added in that
dir and create op comes for that, it does not pass this check below and
we don't process that op.


https://github.com/microsoft/FluidFramework/blob/a51cd2afc8fb11e11c083f9f62fa9e69727e6845/packages/dds/map/src/directory.ts#L2150

Co-authored-by: Jatin Garg <48029724+jatgarg@users.noreply.github.com>
Co-authored-by: Jatin Garg <jatingarg@Jatins-MacBook-Pro-2.local>