feat: Add basic changelog generator tool (#15793)
This tool is used to transform changesets into CHANGELOG.md entries. The
new project is private and in the @fluid-internal scope. It's used as a
`file:` dev dependency in the repo root.

It uses the extensibility that the default changesets tools provide,
which is documented here:
<https://github.com/changesets/changesets/blob/main/docs/modifying-changelog-format.md>

In essence, it takes all the individual changeset files and updates the
CHANGELOG.md file for each package with the changesets that apply to it.
Release engineers will use this when doing minor and major releases.

Unfortunately the changelog APIs are not well documented, so this tool
builds on top of another formatter,
[changesets-format-with-issue-links](https://github.com/spautz/changesets-changelog-format).
The only changes we've made to that formatter is to ignore changelog
entries that are only due to dependency updates. The changelog files
will then be fixed up using the command in PR #15949.

Related to
[AB#3975](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3975)