Add Establishingconnection state when the connection starts happening (#15820)
## Description

Item: https://dev.azure.com/fluidframework/internal/_workitems/edit/3735
Add Establishingconnection state when the connection starts happening.
Currently we remain in Disconnected state until the connection actually
happens successfully(receiving "connect_document_success" and don't
really knows what is happening during the disconnected state.
ConnectionState.EstablishingConnection was added a few months ago but
was not yet implemented. With this PR we implement it internally to tell
us that we actually started establishing the connection. Later we can
also expose this to the app via an event, so that they can also show the
appropriate message to the user if they want to.