Remove chatty telemetry event for client election (#16018)
## Description

The client election logging was initially designed to identify
summarizers running at the same time. There is a problem with this
logging. If we have a document stuck without being able to summarize,
the join/ leave ops the new client processes (catch up) will trigger a
huge number of re-elections. This PR removes the logging.