fix(build-tools): Don't include dynamic version in readmes (#16028)
Including the version numbers in the CLI readme causes issues in CI, and
also means that a build is required after bumping the version, which is
annoying. @alexvy86 has previously noted that the versions in the readme
are not that useful, so I just updated the code to always output 1.0.0.
We could use something else or omit it, but omitting it seems wrong
since it's displayed in the examples using the `--version` flag.
Omitting it when the command explicitly includes the `--version` flag is
misleading and confusing.