Convert webflow example to use esnext modules (#15940)
## Description

This makes progress toward #15917 by converting @fluid-example/webflow
to use esnext modules.

This is the most involved update of any of the packages that depend on
`@fluid-internal/test-version-utils`, since ESNext modules don't support
the `require.extensions` api (they don't support `require`), which was
the technique used to make webpack-style CSS imports not blow up when
run in a node environment. Instead, it's recommended to use the node esm
loader API. Although this API is experimental, it hasn't receive any
changes between v16.12 and the node latest (v20.x).