[GC stress] Remove the FULL test scripts from package.json (#16038)
## Description
After merging with latest main, the CI build is failing due to an error
in how packages are installed for FULL back compat testing. See
https://dev.azure.com/fluidframework/internal/_build/results?buildId=168030&view=logs&j=7e6a3fb7-dbfe-5169-4db8-92b72295ba6c&t=e38f86c1-b224-578d-726f-fe0e17267ef6
for details.
While that is fixed, this is a workaround that removes the scripts that
run full compat tests.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.