Approach for handing off container ref to the migration tool (#16015)
This change adds the ability for the SameContainerMigrator to hand a
reference to the IContainer to the SameContainerMigrationTool. This ref
is expected to be provided by the author of the
ISameContainerMigratableModel (the container author).

We aren't able to get a reference to the IContainer
until the SameContainerMigration tool has been instantiated, so we have
to accommodate late availability with a promise.

With this change, we no longer need the debug buttons for faking a
proposal so they are removed.