Create NodeDidNotRunGC test (#15986)
Rollout Task:
[AB#4370](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4370)
- while trying to rollout the feature, we found this bug.

The goal of this PR is to create a test that throws a NodeDidNotRunGC
error during summarization.

Background:
When rolling out closing the summarizer instead of refreshing v2, which
downloaded the snapshot from the server before closing the summarizer,
telemetry was indicating there was a lot more refreshes than expected.
The ratio was expected to be around 4-6 refresh to 1 close, and it
turned out to be a little less than 2 refreshes to 1 close in preview at
25%. This meant we were exacerbating a problem.

After an investigation, we found that `DataObject`s were being created
during summarization which put the `summarizerNode` system in a
half-baked state. The summary upload would succeed, but the state after
summarization would be inconsistent. After upload we would get a
`NodeDidNotRunGC`. This test intends to mimic a customer scenario. This
does not solve the problem, but we now can verify our solutions.