Draft: Collect memory data for Container Load in production (#14716)
## Description

This is an initial attempt to collect memory data in production to
address
[Task3478](https://dev.azure.com/fluidframework/internal/_workitems/edit/3478)
- The proposed solution described on the task involves using the
experimental API
[measureUserAgentSpecificMemory](https://developer.mozilla.org/en-US/docs/Web/API/Performance/measureUserAgentSpecificMemory)
which requires cross-origin isolation to be enabled on loop pages.
Currently loop has it disabled and we will need to understand the impact
of turning it on before proceeding. Also, the API is async and might
take up to 20 secs to return, which would affect the design of using and
correlating it with our regular events.

Until we have a better understanding of the impact in using it, I looked
at other internal products, and it seems that OWA makes use of the
[performance.memory
API](https://developer.mozilla.org/en-US/docs/Web/API/Performance/memory)
which has been deprecated and will only work on a limited set of
browsers: Ex. Chrome 6+.

I'm releasing as a draft PR to receive some feedback and consider
whether or not it would be beneficial using it for the short-term.