Update SharedTree Node Identifier (Node Key) API (#15844)
## Description

This PR updates/adds the identifier APIs available on `ISharedTreeView`
to match the capabilities of the legacy SharedTree and does the
supporting work/refactors/cleanup necessary for such a change. Node
identifiers are now being referred to as "node keys".

* API generation, conversion, and lookup have been placed under a single
`view.nodeKey` object. This keeps all node key functionality in a single
part of the view API to avoid cluttering the API for users who don't use
node keys.
* Editable tree has a new symbol for directly accessing the local
version of a key for a given node. This makes it similar to the existing
API for getting the stable node key which also uses a symbol. Currently,
retrieving the local key simply gets the stable node key and then
converts it, but in the future it will be changed to be the other way
around.
* The node key types and helpers have been moved into their own folder
in feature-libraries (node-key). These helpers include a comparator for
local node keys since they are typed as `Opaque` (i.e. the user should
not assume they are numbers).
* There is a new interface `NodeKeyManager` which can generate,
stabilize, and localize node keys. A runtime implementation and a mock
implementation for testing are provided.
* The runtime idCompressor has been enabled in TestTreeProvider. This is
currently the only way I know of to get an actual IIDCompressor and so
there are some tests which create a TestTreeProvider for the sole reason
of extracting the IdCompressor out of it (@justus-camp FYI). Note that
for a real SharedTree, the idCompressor must be enabled in the runtime
in order to generate or convert compressed identifiers.

## Breaking Changes

The node identifier map on `SharedTreeView` has moved under the
`nodeKey` object and is now just called `map`.