Adding new matrix + sharedString perf test and adjusting existing tests to improve perf reliability (#16075)
## Description
Re-applying PR
[16044](https://github.com/microsoft/FluidFramework/pull/16044)
Adding a Matrix + Shared String test to the performance mix as the
existing PAS performance test is currently running against the mock
infrastructure. While trying to create alerts for reliability against
the performance benchmark, we've realized that several tests have been
failing for a long time. To illustrate, the 10 Mb summarizer test fails
constantly on FRS and the same goes for several others. For now, we are
decreasing the parameter sizes to improve the overall reliability.
We will gradually add back the same parameters later on and at the same
time investigate each new failure.