Remove IRootSummaryTreeWithStats from ContainerRuntime (#16083)
## Description
`IRootSummaryTreeWithStats` has been removed from container-runtime.
This was added to access `gcStats` from the result of `summarize` in
`submitSummary` method in `ContainerRuntime`. This is no longer the case
after #15663 was merged.

## Breaking Changes
This is a breaking change because it removes exporting
`IRootSummaryTreeWithStats`. However, this was only used within the
`ContainterRuntime` class and no one should depend on this.