feat(build-tools): Add generate changelog command (#15949)
This tool is used to transform changesets into CHANGELOG.md entries. In
essence, it takes all the individual changeset files and updates the
CHANGELOG.md file for each package with the changesets that apply to it.
Release engineers will use this when doing minor and major releases.

Unfortunately the changelog APIs are not well documented, so this tool
builds on top of another formatter,
[changesets-format-with-issue-links](https://github.com/spautz/changesets-changelog-format).
The only changes we've made to that formatter is to ignore changelog
entries that are only due to dependency updates. The changelog files are
then fixed up using a find/replace process by the `generate changelogs`
command.

Related to
[AB#3975](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/3975)

---------

Co-authored-by: Tyler Butler <tylerbu@microsoft.com>