refactor(build-tools): Unify package filter flags across commands (#16054)
## Summary of changes

- Consolidates flags for selecting and filtering packages into two
groups, and adds utility functions to translate flag values into
filter/selection objects and apply those filters to arrays of packages.
- The idea behind selection and filter is that selection is about
_including_ packages, while filtering is about excluding them. That's
not really true, though, because scope is a filter and brings packages
in/out. Maybe a more accurate way to describe it is that there are two
passes - the first pass (selection) includes packages. Then the second
pass (filtering) removes packages that were included in the first pass.
- THis distinction may not be relevant, and an argument can be made that
there should just be filter flags that each are include/exclude. But I
think that's a little more complex overall.
- The release group and release group root filter flags support multiple
values. The release group flag type is _not_ multi-value enabled by
default, so existing commands do not automatically get this behavior.
Commands need to be refactored to use the unified filters to adopt them.
- Adds a `--quiet` flag to the shared command base class, which is the
opposite of the `--verbose` flag. When used, most logging will be
disabled. Output sent using `Command.log` will still be logged. This is
most useful for commands whose text output (as opposed to JSON) is piped
in the shell.
- Adds a `--concurrency` flag to the BasePackageCommand so callers can
control the number of tasks run simultaneously.