ContainerRuntime.closeFn no longer disposes runtime (#15192)
#
[AB#2953](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2953)

`ContainerRuntime` is no longer expected to be disposed upon closing the
container. This needs to be reflected in the `ContainerRuntime.closeFn`
method for consistency.

## Breaking Changes

Calling `ContainerRuntime.closeFn(...)` will no longer call
`ContainerContext.disposeFn(...)` as well. This means the
`ContainerRuntime` will no longer be disposed by calling this method.

To achieve the `ContainerRuntime` being disposed, use the exposed
`ContainerRuntime.disposeFn` method.