Provide different layers' APIs to describeCompat suites (#16108)
## Description

This allows e2e tests which create their own data objects to correctly
test the same compatibility matrix that other e2e tests do. See
https://github.com/microsoft/FluidFramework/pull/12052 for a historical
approach which generally aligns with this one.

Note: this is a reapplication of the changes in #15917. Reverting this
PR didn't fix the issue, and we have since discovered there was a test
reporting issue (fixed in #16098). The intermittent failures are
therefore likely due to a flaky test, and not changes related to the
esnext switch.