feat(build-tools) Add "generate upcoming" command (#16113)
The `generate upcoming` command is used to generate an UPCOMING.md file
from the contents of changesets. The UPCOMING.md file is a replacement
for BREAKING.md.

The command can be run at any time, and will always update the
UPCOMING.md file with the contents of the changesets. The command will
also be run implicitly:

- When a changeset is added using `flub generate changeset`.
- In CI, to ensure it's always up to date with the changesets content.

## Implementation notes

This is a rough implementation, and is mostly string manipulation. I
would prefer a more structured approach using an AST, but tools like
[mdast](https://github.com/syntax-tree/mdast) are ESM-only and
exceedingly difficult to use with the version of TypeScript we have.

Moreover, we don't know exactly what our needs are. Accordingly, I don't
want to spend time fiddling with the output too much yet. Let's see it
in action and let the feedback guide what we do next.