feat(build-tools): Add "list" command to replace "lerna ls" (#16114)
The CI pipeline uses the `lerna ls --toposort` command to enumerate the
list of packages in a release group. This PR adds a new command to the
build tools, `list`, which will replace this use of lerna.

The command also natively supports a `--tarball` flag which outputs the
package tarball names instead of the package names. This eliminates some
shell processing in the CI script.