Add slide parameter to changeInterval listener (#16117)
## Description

The changeInterval event is emitted when slideOnRemove happens. When
customers subscribe to this event to create revertibles for undo-redo,
this event should be ignored as it is not a user edit. But there is
nothing in the information passed to the listener to make this
discernible from user-edited interval moves.

This change adds an additional parameter to the event listener: "slide"
is a boolean that indicates whether the event is a slide caused by
position removal.

I also considered populating the op to the listener (it's not set in
these cases) so it could be checked to see if it was a position remove
or a change interval. But this had quite a large fan out and API
breaking changes (although in APIs that I think we don't really want to
surface externally).

I also considered making a new slide event and splitting that off from
changeInterval, but that felt a bit redundant and is a breaking change
in behaviour (although I don't think there are many users yet).

## Breaking Changes

This adds a new parameter to the listener, but code that doesn't expect
it will continue to work. And code that is expecting it but uses an old
version will get undefined.