Validate summary before uploading it with single commit summaries (#15988)
Currently, the following sequence of events happen during summarize
which can lead to an incorrect summary accepted by the server:
1. Summary is generated.
2. Summary is uploaded to the server. Server accepts this summary and it
becomes the latest summary.
3. completeSummary step fails because of an error such as
"NodeDidNotRunGC".
In step 3, a node was identified as not running GC but the summary was
already uploaded. This summary could have had incorrect GC data.

Move the logic that validates whether a summary had any issues before
uploading the summary to the server. If the validation fails, it is not
uploaded to the server.

This can lead to another issue where if a summary fails because of an
error like "NodeDidNotRunGC", it can keep happening over and over again
because the summarizer will load from the same snapshot and same ops
will be applied which can lead to the same state.
To solve this, if the validate stage fails with "NodeDidNotRunGC" or
"NodeDidNotRunSummary" errors, `retryAfterSeconds` is set in the error
returned to `SummaryGenerator` which will retry summarize with the same
options. Since these errors happens today due to data stores created
during summarize, it should have move past this when the retry happens
and the next summarize should be successful. Added test that
demonstrates this.

[AB#4619](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4619)