Tree: Assign move ID to each moved node (#16091)
## Description

Refactored sequence field to use a scheme where each cell/node moved in
a changeset has its own move ID. A mark which moves multiple nodes will
now have a block of consecutive IDs instead of a single ID representing
the block. This means that when splitting a move mark we no longer need
to allocate new IDs and coordinate the assignment of matching IDs to the
mark representing the other end of the move. To support this
efficiently, `CrossFieldManager`'s interface has been changed to support
reads and writes to ranges of IDs instead of individual IDs.