Explicit dependency injection for ContainerContext (#16153)
This change removes ContainerContext's reference to Container, instead having Container explicitly inject the dependencies that ContainerContext depends upon.  As a result, Container gets to remove, hide, and/or mark readonly a number of its currently-public members.  This change also normalizes some of the interactions (e.g. matching context.setConnectionState() with a context.setAttachState()).