Fluid summary compression (#15656)
This optional/experimental feature uses client-side end-to-end compression to reduce the network and storage costs.  Unlike HTTP compression, blobs remain compressed within the service and therefore end-to-end compression reduces the impact of large documents with respect to storage quota.

Currently this feature uses the LZ4 algorithm, which is typically within ~2x of GZIP while using only 10% of the CPU cycles.  HTTP compression may be used simultaneously for an additional ~2x reduction in networking costs.