[GC stress] Propagate errors correctly through the various layers such that telemetry has the right error (#16177)
## Description
The GC stress test logs an error when an activity such as get attachment
blobs or a data store sending ops fails. However, the actual error that
fails the activity is not always propagated to the parent data store
that logs these errors. This change fixes that by propagating all errors
to the parent data store. The errors are logged from one function now so
it's simpler to make changes and reason about them.

## Reviewer guidance
This is for an experimental stress test in test/gc-stress branch. This
is not in the main FF branch and does not affect its stress test.