Allow rebase in merge tree outside resubmit (#16035)
# Description

The current reconnect logic of the merge tree assumes that all
outstanding operations will be rebased for reconnect, and they will be
done in order of submission. This is always true in the reconnection
case as resbumit is only called after the loader and runtime is sure it
has seen everything it will from the previous client id, as only then is
it safe to resend ops, otherwise we can end up with duplicate ops.

However, if we want to rebase unsent ops that current assumptions break,
specifically that all outstanding ops will be rebased. As now there can
be inflight ops that were generated and sent before the rebase point. To
account for this this change will modify the logic such that on first
rebase, the rebase point is selected, specifically the segment group. We
will take all outstanding segment groups after this rebase point and
stash them. We will enforce that all changes after the rebase point will
be rebased. This is requirement, as any change after the rebase point
becomes invalid when the changes before it have been rebased. There
should also be no new changes interleaved during the rebase process.

related
[AB#4757](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4757)
[AB#4768](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4768)

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>