Fix back compat full config versioning (#16049)
[AB#4657](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4657)

The back-compat with loader pipeline broke because we were getting
`0.0.0-<buildId>-test` from our pkg version, and it was expecting either
`2.0.0-internal.x.y.z` or `2.0.0-dev.x.y.z`. The pipeline should be fine
whether or not -test is appended to the end of the version, i.e
`2.0.0-dev.x.y.z-test`

Not sure what the right fix is here, but these are some options. I'd
like to know if anyone knows how to test this.

---------

Co-authored-by: Tyler Butler <tyler@tylerbutler.com>