Update exception handling logic for uncaughtException (#16003)
## Description

Couple of updates in the runner logic (kafka runner and alfred runner)
to improve exception handling and its logging:
1. uncaughtException events will trigger the runner's promise rejection,
so that it is handled by runService's catch block and exits the process
with exit code 1.
2. The above change will also ensure that "RunService" failure logs
include restarts due to uncaughtExceptions.
3. Added try-catch block in runner's stop() method so that the deferred
promise can be completed before throwing the error.
4. Improved logging and added an 'AlfredRunner' metric in alfred runner.
5. Return alfred runner.stop() early if it has already been stopped.

## Testing
- Tested by manually triggering an uncaught exception in scribe(for
kafka runner) and alfred(for alfred runner), and validated the logs.
- Also tested the other restart scenarios like below and validated the
logs.
    - Exception during a service which is caught by a catch block
    - SIGTERM restart

## Reviewer guidance

These changes will affect the number of restarts captured by RunService
metric, i.e. it will include restarts due to uncaughtExceptions also.
Currently it does not. So we might need to update the alert's
thresholds.

---------

Co-authored-by: Shubhangi Agarwal <shuagarwal@microsoft.com>