tree: PropertyDDS interop package (#15727)
## Description

This is a re-worked version of
https://github.com/dstanesc/FluidFramework/pull/3 (credits go to
@ruiterr, @milanro and @dstanesc).

The new package is aimed to provide tools for a migration from
PropertyDDS to new SharedTree. The first tool is a (runtime) schema
converter. Current implementation is based on the new typed schema API.

_Note that the missing part here for app developers is a way to store
the schema in a file and generate static types, what naturally could be
done by the schema builder (one of many possible approaches)._

## Findings regarding the new schema API

- ~recursive dependencies require special handling (check for a logic
around `cache` variable)~ seems to be the right approach
- ~since builder's `treeSchema` is hidden, one has to use an additional
`treeSchema` map~ - taken over by @noencke
- ~it would be nice to have a native support for generics in the
SharedTree schema API, e.g., for `array` or `map` schemas~ _it would be
even nicer to have a support for generics in general, not only at the
schema level, will come some day..._

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).

The following aspects are to be reviewed:
- new package structure
- overall functionality & tests
- code style vs fluid conventions
- synergy with the typed schema API

The PropertyDDS schema specification-related moments are mostly
clarified, the only two left are (@milanro and @dstanesc , please,
clarify that):
- ~if there are active usecases for nested properties~ - postponed (not
required urgently)
- ~if there is a need to support built-in `RelationshipProperty`~ - this
is unconditionally supported as it inherits from `NodeProperty`, and we
always add all inheritance chain to the schema.

~What definitely requires improvement is a README, but this I'll work on
currently.~

---------

Co-authored-by: Roland Ruiters-Christou <roland.ruiters@autodesk.com>