Enhance reliability of obtaining colletion from interval (#16251)
## Description


[AB#4524](https://dev.azure.com/fluidframework/internal/_workitems/edit/4524),
[AB#4081](https://dev.azure.com/fluidframework/internal/_workitems/edit/4081)

This PR implements enhancements to improve the reliability of obtaining
interval collections from intervals. Following discussions with
@scarlettjlee, we concluded that directly making the properties private
is not feasible since we need to ensure the existing data at rest
continue to work. As a result, several compromise strategies are
included in this PR:

1. Prevent overwriting the collection label when adding intervals or
explicitly changing properties, it is for the purposes:

- Each interval should only ever be in one collection, their labels
should be consistent

- We encourage users to create/add an interval based on existing
collection, rather than creating random interval and insert them into
collections

~~2. Add a new API to `ISerializableInterval` to get the labels,
allowing users to retrieve the labels similar to `getIntervalId()`. We
discourage users to get access to the properties directly, despite they
are public.~~