offline: fix redirect table blob not being saved and re-enable stashed ops in stress test (#16031)
#15831 caused containers with offline load enabled to download all
attachment blobs in the snapshot, resulting in throttling in the stress
test. Also, it caused an assert to fire when attaching a detached
container with blobs with offline load enabled.

This PR saves only the redirect table from the .blobs tree, adds a
regression test for attaching with blobs with offline load enabled, and
re-enables stashed ops in the stress test.