Handle deleted/revived subtrees and deactivate/reactivate paired marks (#16235)
## Description
Before this change, we would crash when the ancestor of a node being
moved was concurrently deleted. This fixes that crash and gives the
desired merge semantics (the move doesn't occur unless the ancestor is
revived).

## Reviewer Guidance

The review process is outlined on [this wiki
page](https://github.com/microsoft/FluidFramework/wiki/PR-Guidelines#guidelines).