Dispose summarizer instead of closing (#16143)
For summarizer, there is no user to call `Container.dispose(...)`. So,
if code ever closes the summarizer, it should dispose instead.

Where we used to have a single container shutdown concept ("close"), now
we also have "dispose". The difference being:
- "close" is how an app or internal code should respond to a critical
error, assuming the user is still there. This gives the user the chance
to review the document read-only before leaving the session entirely
- "dispose" is for completely tearing down the container and all
resources, for when the user is gone.
For the Summarizer, there is no user. If the code ever closes it, it
should immediately be disposed.


[AB#4610](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4610)