fix: Correct handling of filter and selectionFlags (#16254)
## Description

The `--dir` flag in the `exec` command is not working, and the root
cause is that the code expected the flag to be called `directory` but it
was called `dir`. The reason this wasn't caught was because of sloppy
typing. I had used an `any` to parse the raw flags into filter and
selection objects, which introduced this error.

## Fix

The flag types are somewhat complex, so it seemed simpler to make
simpler "mirror types" that structurally match the oclif flag types, but
are typed more simply. This allows me to add some type safety to the
parse functions.

## Testing

I added several new tests that exercise the selection and filtering.
Most tests work against the helper filtering functions. However, there
are also tests that run through a new test-only command to validate that
converting oclif flags into filter/selection objects is done correctly.