Remove unnecessary/missed common definitions imports and deps (#16281)
This is a follow-up to
https://github.com/microsoft/FluidFramework/pull/16211, removing the dep
`@fluidframework/common-definitions` if its not used in the package and
catch any stray ones I may have missed

The type test infrastructure seems to sometimes output type tests for
re-exported APIs, but not always. Since the type test changes in this PR
are related to that - a re-exported API was tested before and no longer
is - the change should be non-breaking.