fix(protocol-base): Correct spelling of buildGitTreeHierarchy (#16323)
The buildGitTreeHierarchy function name was misspelled, so I corrected
it. We haven't yet released any version with the misspelled name (though
there is a prerelease version that contains it), so it's safe to just
correct it. No need to export the old wrong name for backwards compat.

I also corrected some similar misspellings in other unexported APIs and
in some comments.