Submit blobAttach ops in separate un-grouped batch (#15794)
## Description

`"blobAttach"` ops need their metadata to be visible at the service
level, so they cannot be included as part of the grouped batching
process.
Thus, we need to have them in their own, separate batch from normal ops
(which will go through grouped batching).

**Note:** This hinges on the assumption that the order of `"blobAttach"`
ops doesn't matter too much as long as they are submitted before the
blob handles are used in regular ops.


[AB#4221](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/4221)

## Note about previous implementation

Previously, along with the introduction of the grouped batching feature,
all `"blobAttach"` ops were sent in their own batch as well. See
[#14512](https://github.com/microsoft/FluidFramework/pull/14512/files#diff-d6dd5b7a4910beea419ad4e2e60191c88283bac177f1d87b9046be1592a763feR1211).

However, this was done inside a `Promise.resolve(...)` where it would
flush the current batch, submit the `"blobAttach"` op, then flush that
batch. This was a problem because runtime layers would think it already
submitted the `"blobAttach"` op when referencing the blob handle, and
ops referencing the blob handle could appear earlier in the op stream
because of the `Promise.resolve(...)` wrapper.
This wrapper couldn't be removed as we could currently be batching when
the `"blobAttach"` op is sent, and batch semantics would be broken.

Temporary solution at the time was to simply prevent grouped batching on
any batch containing a `"blobAttach"` op.