Revert "Update deps (newest) on server release group in the client release group" (#16337)
Reverts microsoft/FluidFramework#16311

This change is causing execution time breaks in our real service tests
due to a breaking change in a semver compatible release of server. we
need to fix the release, or update the next branch. for now we will
revert the bump while the real fix is determined.