Update branching/transaction APIs (#16319)
## Description

This makes some changes to the branching API, how it interacts with
transactions, and related APIs. Namely:

* Branches that merge into other branches are now automatically
disposed, unless explicitly told not to. If the target branch is in the
middle of a transaction, the branch being merged in _must_ be
automatically disposed.
* Merging a branch which is disposed now fails.
* If a branch is in the middle of a transaction and is merged into
another branch, its transaction(s) is/are automatically committed.
* `SharedTreeBranch`'s `dispose` method can now be called multiple times
(second calls have no effect).
 
This PR also fixes a bug in trunk eviction and adds a helper for
registering listeners on a branch and all of its descendants.

An additional bug was discovered regarding branch merging during the
detached phase, and a bug has been filed.

## Breaking Changes

Branches are now automatically disposed when merged; to prevent this and
maintain the previous behavior, pass `false` as the second argument to
the `merge` method.