Fix the inconsistency between local delete and remote clear (#16340)
[AB#4612](https://dev.azure.com/fluidframework/internal/_workitems/edit/4612)

When encountering a remote clear operation, `clearExceptPendingKeys`
will be called to retain the local modified but unack'd keys, however,
it just treats all unack'd operations as `set`. For instance, if a local
delete `delete("key1")` coincides with the remote clear, it will clear
the map first, then force a `set("key1", undefined)`. It is necessary to
determine if the unack'd operation is `delete` or `set` operation, this
can be achieved by checking if the affected key still exists in the data
or not.

---------

Co-authored-by: Abram Sanderson <Abram.sanderson@gmail.com>